Extract touchpad logic into new TouchpadHandler class. Use Settings parameter LongpressTimeout for debounce.
